Support for heal_instance_info_cache_interval
Nova caches a copy of the instance's network information by reaching
to neutron for network data. This action will be executed on every
heal_instance_info_cache_interval (default to 1min). In large scale
deployment controlling this param will reduce the load on the
neutron-server api and keystone.

The original work on this was done by Liran Cohen of LivePerson; the
finalization of the patch was done by Martin Loschwitz from SysEleven.

Change-Id: Ie3a09a03e7b831fca22b2d6275e3c611b020a39f